Cprowesl adds title and Window Size functions
I made it so that we can see the buttons via the configuration dialogue.  I added a relevant title to the window.  I changed some window sizing options so that the window appears as larger and gives more room for work.  I may be changing the size of the buttons soon.