Cprowesl pushes button formatting and font changes.
I updated the fonts and colors on the main menu.  This is a temporary change but it made the buttons more legible for us to work with.  Warning: I have noticed some attributes of the buttons do not work.  At least, it seems like they have no effect on my end.  Many of tkinters functions seem to be deprecated in some areas.  This will be my push before I complete task SMIP #15: Make GUI Frame.