Cprowesl adjusts Lesson Page layout
The lesson page now has multiple frames which make up each lessons "drawing board".  The frames are {header, slide, ribbon} so far.  Each lesson has a header which appears at the top left with a very brief title for the lesson.  The header is followed by an instruction for the student to follow.  We should start keeping a dictionary or file of these somewhere.  Then comes an entry box and a submit button.  Below all this is the ribbon frame which holds the main menu return button.  I have been having some issues with the sizing of the ribbon so exercise caution.  The user can enter text into the entry box and hit submit which will call the subroutine 'feedback'.  At the moment feedback wipes slide clean and then reinserts the entry box, submit button, and a brand new feedback label.  The feedback label is meant to show what their last entry was and if it was correct or incorrect.  We still need the application to send this data to spim and return a positive or negative response.  We still need to get the main menu return button working.