Fixed minor merge issue, updated gitignore
Fixed issue in the merge that had the old label still referenced even though it is no longer used, was over looked in the merge conflict and caused errors at runtime but has since been fixed. Updated gitignore to ignore and stop tracking more of the binary and other autogenerated files.