Cprowesl changes Lesson Layout
The lesson page now has multiple frames which make up each lessons "drawing board".  The frames are {header, slide, ribbon} so far.  Each lesson has a header which appears at the top left with a very brief title for the lesson.  The header is followed by an instruction for the student to follow.  We should start keeping a dictionary or file of these somewhere.  Then comes an entry box, a hint button, and a submit button.  Notice I moved the Hint Button code around so that it can be called in the feedback subroutine.  Below all this is the ribbon frame which holds the main menu return button.  I have been having some issues with the sizing of the ribbon so exercise caution.  The user can enter text into the entry box and hit submit which will call the subroutine 'feedback'.  At the moment feedback wipes slide clean and then reinserts the entry box, hint button, submit button, and a brand new feedback label.  In the future we may want to give the feedback subroutine a dictionary of buttons to save on parameter space.  The feedback label is meant to show what their last entry was and if it was correct or incorrect.  We still need the application to send this data to spim and return a positive or negative response.  Gursimrat Singh will be providing code to output it to a file.  We still need to get the main menu return button working.  The Style name of the page was irrelevant and needed to be renamed to fit the menu theme for Linux users.  I changed it from black/green to B_DO1 in order to represent the combination of the foreground and background colors, which are Black and DarkOrange1.  We should adjust this scheme next week so that themes are only defined once and proliferate the app.