Cprowesl Adds Hint Popup
I added functionality to the hint button so that when you select it, a dialogue box appears with a sample hint for the user.  The box hovers over the user entry area so that it grabs their attention.  They can click 'OK' or hit <Enter> to exit the dialogue.  The cursor setting was changed to the target symbol for buttons within the lesson page to give it an interactive and uniform look.  Some spacing was changed to benefit PEP8 standards.