Added a Drop Down for References Button inside a Lesson.
When the reference button in a lesson is clicked, it shows a pop up menu with 2 options: Green Sheet and Test Link. More options can be added to each individual lessons. Changed the design of how references are stored. The reference button on main menu still works like it used to.