Cprowesl restyles Create Lessons v1
NOTE: I was moved to this task following instruction from team.  My other task is to implement the previous lesson button.  In this revision, I refactored the names of some of the styles, as well as added new ones.  I made a small change to Main.py which reflects our need to keep the window as tall as possible for our buttons since we don't have a scroll bar.  In Drawer.py, the bulk of my changes occurred.  When we draw a main menu, a check is done to see if the window size has changed.  If the window has changed size, it adjusts accordingly.  Some settings were changed to maximize the surface area of the frame and material.  Due to sizing issues the register label strings are now adjusted and sliced as they are processed.  New fonts and styles have been applied.  There is now a Reference label with a drop down menu and an 'Add a Reference' button.  I tried to use as much of the space as I could.  The screen needed more space for the rows so I expanded the border by 200.  This screen is only hooked up to the Main Menu through a button.  Therefore, I handled the change in window size when it snaps back to the menu.