Preemptive Bug Fix with updating lesson list
When a new lesson is created in the create lessons form that new lesson object is appended to the end of the current lessons list. However the size of that lessons list was assigned to a variable at the launch of the program and was not updated. Now we just get the immediate length of the list.