Cprowesl Adds Create Lesson Feedback
In this push I created a label which will pop up over the create lesson button once it has been pressed and processes the user's request.  On Linux the color of the label is "green2".  This may vary based on OS.  I had to import sleep from time for this to work.  I provided the user with a three second window to see that the Lesson has been created before I grid_forget() the label.  The label is forgotten but NOT destroyed at this point.  Take that into consideration with modifications.  I also changed the text entry on the submit_lesson_button to a tk.StringVar().  It may not be necessary to keep this as the project comes to a close, so be aware.  I padded the y var of the Main Menu and Create Lesson buttons to make the frame look more spacious.