Added validation checks on lesson creation form to prevent issues
Now when the lesson creation form is submitted it checks that the values in the forms are valid before it creates the file and prompts the client accordingly. Also modified the way the lesson created popup works. Added a popup if an issue is caught. Made label clear with a call to after rather than to sleep