Cprowesl creates Lesson Transitions
In this push, I meddled with lesson transitions through the transfer_to function, and get_next/previous_lesson functions.  This included the painstaking gridifying of the slide.  Some functionality that is leftover from .pack() methods are inert and should be removed.  This will be saved for Sprint 3 revisions.  We still need a get_current_lesson function to be coded.  Any questions, message me