Added coloring of lesson_select based on completion of a lesson
Currently a WIP, It reads and writes to a .pickle file to remember what lessons have been completed throughout different sessions, so lesson_select can color code lessons with green = completed, red = Submitted but wrong, yellow = not started.

Does not currently work, I'm going to discuss it with everyone and change it tomorrow to be done in a different way preferably, because this current way causes too many headaches for what it's worth imo.