Cprowesl pushes start of GUI unit tests
In this push, I am pushing the beginnings of the GUI unit tests we will be performing for Sprint 2.  At this moment, we are passing 2 of 2 available tests.  These tests use a tkinter GUI Label to test that a Label can be made, and then test that the text within a GUI Label can be configured after it has been spawned.  A link has been included at the top of the file to show the example GUI tests we found available on the internet.  Our code is going to mirror this approach and try to make tests that SMIP specific by only including tests for GUI elements that are available in the source code of SMIP.