First back end unit test
I currently include all these excess file python generates, I'm not sure if we need all this garage in here. 

I've pushed a hardcoded change to Lesson_Workbook, we need to change the gui testing direcotry to match this one so they both work. If you try to run the front end tests on this branch, they will not work.