Bug Fix: Lesson Creation Validation
When doing last minute testing found a bug in the way we validate inputs on the lesson creation form. Would not catch cases where invalid characters followed numbers. Changed implementation to check that all characters are numbers. This bug managed to get through all previous tests we did and that is why the fix is happening here.