Added grading functionality, including GUI Page and excel output
Added a Grading button to the main menu that when clicked take you to the grading page. The client then selects a lesson to grade and the directory for all the submissions. When 'Grade' is clicked the program grades all the submissions and outputs the results to a spreadsheet in SMIP/grading