Cprowesl completes lesson visual feedback
This push is in regards to issues #78 and #79.  A green/red label now appears when a submitted lesson is correct/incorrect.  The green label reads 'PASSED', while the red label reads 'FAILED'.  Some extra white space characters are used in the labels to provide button coverage.  I had to edit the submit_code() function where it is declared, and where it appears.  For example, I was required to change the parameter list in the draw_practice() function so that it reflected my changes to submit_code().  Let me know if any other side effects went unaccounted for.